--- src/components/BookDetail/styles.css
@@ -1,37 +1,0 @@
-.book-container {
-  display: flex;
-  align-items: center;
-  padding: 2rem;
-  box-shadow: 1px 1px 3px #ababab;
-  border-radius: 0.5rem;
-}
-
-.book-container + .book-container {
-  margin-top: 1rem;
-}
-
-.book-container .book-infos {
-  flex: 1;
-  margin: 0 1rem;
-}
-
-.book-infos .book-title {
-  font-size: 1.5rem;
-  font-weight: 500;
-  color: var(--gray-900);
-  margin-bottom: 0.5rem;
-}
-
-.book-infos .book-description {
-  font-size: 1rem;
-  font-weight: 400;
-  color: var(--gray-500);
-  margin-bottom: 0.5rem;
-}
-
-.book-infos .book-author {
-  font-size: 0.875rem;
-  font-weight: 400;
-  color: var(--gray-400);
-  margin-top: 1rem;
-}
