--- src/components/Form/styles.css
@@ -1,104 +1,0 @@
-*{
-    
-    margin: 0;
-    padding: 0;
-    font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
-}
-/* Image */
-.screen .image{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    
-}
-
-.screen .image img{
-    width: 70%;
-}
-
-.screen .image h2{
-    margin-top: 60px;
-    font-size: 50px;
-    color: #555;
-
-}
-
-.screen{
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    height: 100vh;
-    
-}
-
-.screen section{ 
-    background-color: rgb(255, 255, 255);
-    width: 50%;
-    height: 100%;
-}
-
-/* Form */
-
-.FormS{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    background-image: radial-gradient( circle farthest-corner at 10% 20%,  #61baff 0%, #a6effd 90.1% );
-}
-
-.FormS h2{
-    margin: 30px 0px 30px 0px;
-    font-size: 35px;
-    color: rgb(44, 44, 44);
-    font-weight: 800;
-}
-
-.FormS label{
-    font-size: 20px;
-    margin-right: 20px;
-    font-weight: 700;
-    
-}
-
-.FormS input{
-    display: flex;
-    flex-direction: row;
-    align-items: flex-end;
-    border-top-style: none;
-    border-right-style: none;
-    border-left-style: none;
-    background-color: transparent;
-    border-bottom: 1px solid #35a1f3;
-}
-
-input:focus {
-	outline: none;
-  	border: none;
-    border-bottom: 1px solid #a600dd;
-
-}
-
-form{
-    width: 70%;
-    min-height: 70%;
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-
-}
-
-form button{
-    width: 70%;
-    border-radius: 5px;
-    padding: 10px;
-    margin: 20px;
-    border: none;
-    font-size: 20px;
-    color: #212121;
-    cursor: pointer;
-    background-color: #fff;
-}
