+++ src/components/Header/index.jsx
@@ -2,7 +2,7 @@ import logo from "../../assets/images/logo.png";
 import "./styles.css";
 import React from "react";
 
-import Nav from "../../Nav";
+import Nav from "../Menu/Nav";
 
 export const Header = () => (
   <header>
