--- src/pages/Form/index.js
@@ -1,12 +1,0 @@
-import React from "react";
-import './styles.css';
-
-import Form from "../../components/Form";
-
-export default function NewBook() {
-    return (
-      <div className="newbook-container">
-        <Form />
-      </div>
-    );
-};
