--- src/pages/Form/styles.css
@@ -1,8 +1,0 @@
-.newbook-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  font-family: Arial, Helvetica, sans-serif;
-  padding-top: 100px;
-}
