+++ src/pages/Livro/BookDetail.jsx
@@ -1,8 +1,7 @@
+import React from 'react'
 import "./styles.css";
 import { AiFillHeart, AiOutlineHeart } from "react-icons/ai";
 import { ImBooks } from "react-icons/im";
-import React from 'react'
-import React from "react";
 
 export const BookDetail = ({ book }) => (
   <div className="book-container">
